[Package] [Handlers] Cleanup Code Base
- [+] feat(handlers/init.go): remove unused imports and variables
- [+] feat(handlers/internal.go): remove unused imports
- [+] feat(handlers/logger.go): add logger variable and functions
- [+] feat(handlers/payload.go): add payload structs and bindUpdatePayload function
- [+] feat(handlers/url.go): add handlers for getting, posting, editing, and deleting URLs
